requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
pendulum
python-dotenv
